+++ pet_shop/src/main/java/com/example/pet_shop/service/SubcategoryService.java
@@ -1,5 +1,6 @@
 package com.example.pet_shop.service;
 
+import com.example.pet_shop.exceptions.BadRequestException;
 import com.example.pet_shop.model.DTOS.SubcategoryDTO;
 import com.example.pet_shop.model.entities.Category;
 import com.example.pet_shop.model.entities.Product;
@@ -13,23 +14,29 @@ import java.util.Optional;
 @Service
 public class SubcategoryService extends AbstractService{
 
-    public Subcategory createSubcategory(SubcategoryDTO subcategoryDto) {
+    public Subcategory createSubcategory(SubcategoryDTO subcategoryDto,int cId) {
         Subcategory subcategory = new Subcategory();
         subcategory.setName(subcategoryDto.getName());
 
-        Optional<Category> optionalCategory = categoryRepository.findById(subcategoryDto.getCategoryId());
+        Optional<Category> optionalCategory = categoryRepository.findById(cId);
         if (optionalCategory.isEmpty()) {
-            throw new EntityNotFoundException("Category with ID " + subcategoryDto.getCategoryId() + " not found");
+            throw new EntityNotFoundException("Category with ID " + cId + " not found");
         }
+
+        optionalCategory.get().getSubcategories().add(subcategory);
         subcategory.setCategory(optionalCategory.get());
         return subcategoryRepository.save(subcategory);
     }
 
-    public void deleteSubcategory(Integer subcategoryId) {
+    public void deleteSubcategory(int subcategoryId) {
+        if(!subcategoryRepository.existsById(subcategoryId)){
+            throw new BadRequestException("No such subcategory.");
+        }
         subcategoryRepository.deleteById(subcategoryId);
     }
 
+
     public Page<Product> getProductsBySubcategoryId(int subcategoryId, Pageable pageable) {
-        return productRepository.findAllBySubcategoryId(subcategoryId,pageable);
+        return subcategoryRepository.findAllProductsBySubcategoryId(subcategoryId, pageable);
     }
 }
